installing via nasher forces yes instead of no in case the compiler doesn't work, added a safety check to the production script to stop if no module was generated (untested)

+++ linux_run_server.sh
@@ -21,6 +21,11 @@ Last Updated: $timestamp ($hash)"
 
 $PWD/tools/linux/nasher/nasher install --erfUtil:"$PWD/tools/linux/neverwinter/nwn_erf" --gffUtil:"$PWD/tools/linux/neverwinter/nwn_gff" --tlkUtil:"$PWD/tools/linux/neverwinter/nwn_tlk" --nssCompiler:"$PWD/tools/linux/nwnsc/nwnsc" --installDir:"$PWD" --nssFlags:"-oe -i $PWD/nwn-base-scripts" --no --modDescription="$desc"
 
+if [[ ! -f TFN.mod ]] ; then
+    echo 'Module does not exist, aborting.'
+    exit
+fi
+
 # rm server/config/common.env
 rm server/modules/TFN.mod
 rm -d -R  server/override
